+++ sonaranalyzer-dotnet/src/SonarAnalyzer.VisualBasic/Metrics/Metrics.cs
@@ -41,7 +41,7 @@ namespace SonarAnalyzer.Common.VisualBasic
                 throw new ArgumentException(InitalizationErrorTextPattern, nameof(tree));
             }
 
-            publicApis = new Lazy<ImmutableArray<SyntaxNode>>(GetPublicApiFactory(tree.GetRoot()));
+            publicApis = new Lazy<ImmutableArray<SyntaxNode>>(() => PublicApiMetric.GetMembers(tree));
         }
 
         protected override bool IsEndOfFile(SyntaxToken token) =>
@@ -101,14 +101,6 @@ namespace SonarAnalyzer.Common.VisualBasic
 
         protected override IEnumerable<SyntaxNode> PublicApiNodes =>
             publicApis.Value;
-
-        private static Func<ImmutableArray<SyntaxNode>> GetPublicApiFactory(SyntaxNode root) =>
-            () =>
-            {
-                var walker = new PublicApiWalker();
-                walker.Visit(root);
-                return walker.PublicApi;
-            };
 
         private bool IsComplexityIncreasingKind(SyntaxNode node)
         {
